US32

+++ app/Http/Controllers/AeronaveController.php
@@ -60,22 +60,30 @@ class AeronaveController extends Controller
 	}
 
     public function pilotosAutorizados(Aeronave $aeronave){
+    	$this->authorize('pilotosAutorizados', Aeronave::class);
+
         $pilotos = Aeronave::find($aeronave->matricula)->pilotos()->paginate(15);
         return view('aeronaves.listagemPilotosAutorizados',compact('pilotos','aeronave'));
     }
 
     public function naoAutorizarPiloto(Aeronave $aeronave, User $piloto){
+    	$this->authorize('pilotosAutorizados', Aeronave::class);
+
         $aeronave->pilotos()->detach($piloto->id);
         return redirect()->route('aeronaves.pilotosAutorizados',$aeronave)->with('sucesso', 'Piloto adicionado à lista de não autorizados!');
     }
 
     public function pilotosNaoAutorizados(Aeronave $aeronave){
+    	$this->authorize('pilotosAutorizados', Aeronave::class);
+
     	$todosPilotos= $aeronave->pilotos()->get()->pluck('id')->toArray();
     	$pilotos = User::where('tipo_socio','P')->whereNotIn('id',$todosPilotos)->orderBy('id')->paginate(15);
         return view('aeronaves.listagemPilotosNaoAutorizados',compact('pilotos','aeronave'));
     }
 
     public function autorizarPiloto(Aeronave $aeronave, User $piloto){
+    	$this->authorize('pilotosAutorizados', Aeronave::class);
+    	
         $aeronave->pilotos()->attach($piloto->id);
         return redirect()->route('aeronaves.pilotosAutorizados',$aeronave)->with('sucesso', 'Piloto adicionado à lista de autorizados!');
     }
